+++ init.rb
@@ -2,16 +2,19 @@ Redmine::Plugin.register :entra_id do
   name "Entra ID"
   author "EEA"
   description "Enable login using Microsoft EntraID"
-  version '1.0.0'
+  version '1.1.0'
   url "https://github.com/eea/redmine_entra_id"
   author_url "https://github.com/eea/redmine_entra_id/graphs/contributors"
 
   settings default: {
     enabled: false,
     exclusive: false,
+    auth_method: 'secret',
     client_id: "",
     client_secret: "",
-    tenant_id: ""
+    tenant_id: "",
+    certificate_path: "",
+    certificate_password: ""
   }, partial: "settings/entra_id"
 end